add board user to mypage responseDto
