+++ build.gradle
@@ -31,6 +31,8 @@ dependencies {
     implementation group: 'com.auth0', name: 'java-jwt', version: '3.10.3'
     implementation 'io.jsonwebtoken:jjwt:0.9.1'
     implementation 'org.junit.jupiter:junit-jupiter:5.7.0'
+    implementation 'org.projectlombok:lombok:1.18.18'
+    implementation 'org.projectlombok:lombok:1.18.18'
     // https://mvnrepository.com/artifact/com.h2database/h2
 //    compile('com.h2database:h2')
     compile 'org.json:json:20210307'
