+++ src/main/java/com/sparta/hanghae/picturespot/dto/response/mypage/MypageResponseDto.java
@@ -2,6 +2,7 @@ package com.sparta.hanghae.picturespot.dto.response.mypage;
 
 import com.sparta.hanghae.picturespot.dto.request.img.BoardImgCommonRequestDto;
 import com.sparta.hanghae.picturespot.model.Board;
+import com.sparta.hanghae.picturespot.model.User;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -18,7 +19,9 @@ public class MypageResponseDto {
 
     //board
     private Long boardId;
+    private Long writerId;
     private String writer;
+    private String writerImgUrl;
     private String title;
     private String content;
     private double latitude;
@@ -33,15 +36,17 @@ public class MypageResponseDto {
 
     //comments
     private List<MypageCommentResponseDto> comments;
-    private List<BoardImgCommonRequestDto> boardImgReponseDtoList = new ArrayList<>();
+    private List<BoardImgCommonRequestDto> boardImgResponseDtoList = new ArrayList<>();
 
 
     @Builder
-    public MypageResponseDto(Board boardEntity, List<MypageCommentResponseDto> comments, boolean likeCheck, int likeCount, List<BoardImgCommonRequestDto> reponseDto) {
+    public MypageResponseDto(Board boardEntity, List<MypageCommentResponseDto> comments, boolean likeCheck, int likeCount, List<BoardImgCommonRequestDto> responseDto) {
 
         //board 정보
         this.boardId = boardEntity.getId();
+        this.writerId = boardEntity.getUser().getId();
         this.writer = boardEntity.getUser().getNickname();
+        this.writerImgUrl = boardEntity.getUser().getImgUrl();
         this.title = boardEntity.getTitle();
         this.content = boardEntity.getContent();
         this.category = boardEntity.getCategory();
@@ -50,7 +55,7 @@ public class MypageResponseDto {
         this.spotName = boardEntity.getSpotName();
         this.modified = boardEntity.getModified();
         //이미지
-        this.boardImgReponseDtoList = reponseDto;
+        this.boardImgResponseDtoList = responseDto;
 
         //좋아요
         this.liked = likeCheck;
