+++ src/test/resources/application.yml
@@ -1,0 +1,83 @@
+spring:
+  security:
+    oauth2:
+      resource:
+        jwt:
+          key-uri: http://localhost:8080/oauth/token_key
+      client:
+        registration:
+          google:
+            client-id: 197445114533-d38vg2t645888355rt20mig52gqjo1bo.apps.googleusercontent.com
+            client-secret: PGiKidfkboA-r8EQHiKoEI7o
+            redirect-uri: http://13.125.97.117/login/oauth2/code/google
+            scope: email, profile
+          naver:
+            client-id: qkQkFErqozdauGrjUOQ3
+            client-secret: IWvQLImorG
+            scope: name, email, profile_image
+            client-name: Naver
+            authorization_grant_type: authorization_code
+            redirect-uri: http://13.125.97.117/login/oauth2/code/naver
+          kakao:
+            client-id: e874ec34648c4f9d1174e502e3f33ea2
+            #client-secret: CEAI59T1UVmb9WTptGT6jBS63lETtOgi
+            scope: profile, account_email
+            client-name: Kakao
+            authorization_grant_type: authorization_code
+            redirect-uri: http://13.125.97.117/login/oauth2/code/kakao
+        # provider 네이버는 oauth2에 provider가 아니기 때문에 설정해줘야함 (구글, 페이스북 같은 거는 등록되있다.)
+        provider:
+          naver:
+            authorization-uri: https://nid.naver.com/oauth2.0/authorize
+            token-uri: https://nid.naver.com/oauth2.0/token
+            user-info-uri: https://openapi.naver.com/v1/nid/me
+            user-name-attribute: response # 회원정보를 json으로 받는데 response라는 키값으로 네이버가 리턴해줌
+          kakao:
+            authorization-uri: https://kauth.kakao.com/oauth/authorize
+            token-uri: https://kauth.kakao.com/oauth/token
+            user-info-uri: https://kapi.kakao.com/v2/user/me
+            user-name-attribute: properties
+  app:
+    auth:
+      tokenSecret: 926D96C90030DD58429D2751AC1BDBBC
+      tokenExpirationMsec: 864000000
+    oauth2:
+      # OAuth2 공급자로 성공적으로 인증 한 후 사용자에 대한 인증 토큰을 생성하고 토큰을
+      # 프론트 엔드 클라이언트가 /oauth2/authorize 요청에서 지정한 redirectUri입니다.
+      # 쿠키는 모바일 클라이언트에서 잘 작동하지 않기 때문에 사용하지 않습니다.
+      authorizedRedirectUris:
+        - http://localhost:3000/oauth2/redirect
+  servlet:
+    multipart:
+      max-request-size: 10MB
+      max-file-size: 10MB
+  mail:
+    host: smtp.naver.com
+    port: 465
+    username: sflashservices@gmail.com
+    password: tmvmffotnltjqltm!
+    properties:
+      mail.smtp.auth: true
+      mail.smtp.ssl.enable: true
+  datasource:
+    url: jdbc:mysql://springboot-db.cqgiaebdipjp.ap-northeast-2.rds.amazonaws.com:3306/sflash?useSSL=false&useUnicode=true&characterEncoding=utf8
+    username: admin
+    password: jj05161217jj
+  jpa:
+    hibernate:
+      ddl-auto: update
+    properties:
+      hibernate:
+        show_sql: true
+        dialect: org.hibernate.dialect.MySQL5InnoDBDialect
+cloud:
+  aws:
+    s3:
+      bucket: picturespot
+    region:
+      static: ap-northeast-2
+    stack:
+      auto: false
+    credentials:
+      accessKey: AKIA4PAY5UU4VMHS4TMJ
+      secretKey: 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
